requirements.txt
fastapi
uvicorn
spotify-dl
pysondb-v2
py-deezer
fastapi-socketio